// src/store/type.ts
export interface UserInfo {
  uid: string;
  nick: string;
  group: string;
  token: string;
}

export interface Employee {
  key: string;
  realName: string;
  authorEmail: string;
  authorName: string;
  userEmail: string;
  userName: string;
  phone: string;
}

export interface Signkey {
  key: string;
  token: string;
}